requirements.txt
dbt-duckdb==1.9.1
duckdb==1.2.1
pandas==2.2.3
numpy==2.2.4
pydantic==2.10.6
langchain==0.3.19
langchain-openai==0.3.9
openai==1.75.0
streamlit==1.43.2
plotly==5.24.1
folium==0.19.3
streamlit-folium==0.23.2
python-dotenv==1.0.1
scipy==1.15.2
requests==2.32.3
